CreationEngineSource/electron-app/output/south park/requirements.txt
manim
gtts
nltk
fastapi
uvicorn
tkinter
moviepy
Pillow
python-dotenv
gTTS